# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bon Barang Lapas API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  pendingCount: number;
  approvedThisMonth: number;
  totalRequests: number;
  fulfilledCount: number;
}