# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bon Barang Lapas API
 * OpenAPI spec version: 0.1.0
 */

export interface Item {
  id: number;
  name: string;
  defaultUnit: string;
  category: string;
  /** @nullable */
  currentStock?: number | null;
}