# lib/api-zod/src/generated/types/itemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bon Barang Lapas API
 * OpenAPI spec version: 0.1.0
 */

export interface ItemUpdate {
  name?: string;
  defaultUnit?: string;
  category?: string;
  /** @nullable */
  currentStock?: number | null;
}