# lib/api-zod/src/generated/types/parseResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bon Barang Lapas API
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedItem } from './parsedItem';

export interface ParseResult {
  items: ParsedItem[];
}