# lib/api-zod/src/generated/types/requestDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bon Barang Lapas API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalHistory } from './approvalHistory';
import type { RequestDetailStatus } from './requestDetailStatus';
import type { RequestItem } from './requestItem';

export interface RequestDetail {
  id: number;
  requesterId: number;
  requesterName?: string;
  requesterSection?: string;
  status: RequestDetailStatus;
  createdAt: string;
  location: string;
  requestDate: string;
  /** @nullable */
  approvedBy?: number | null;
  /** @nullable */
  approvedByName?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  items: RequestItem[];
  approvalHistory: ApprovalHistory[];
}