# lib/api-zod/src/generated/types/requestItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bon Barang Lapas API
 * OpenAPI spec version: 0.1.0
 */

export interface RequestItemInput {
  itemName: string;
  quantity: string;
  unit: string;
  purpose: string;
  itemNo: number;
}