# lib/api-zod/src/generated/types/requestSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bon Barang Lapas API
 * OpenAPI spec version: 0.1.0
 */

export interface RequestSummary {
  id: number;
  requesterName: string;
  status: string;
  createdAt: string;
  itemCount: number;
}