# lib/api-zod/src/generated/types/topItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bon Barang Lapas API
 * OpenAPI spec version: 0.1.0
 */

export interface TopItem {
  itemName: string;
  requestCount: number;
  totalQty: string;
}